Move helper functions into new module utils.py
