+++ default/python/AI/test/utils.py
@@ -1,0 +1,13 @@
+import unittest
+
+
+def greater_or_equal(arg):
+    return unittest.TestCase.assertGreaterEqual, arg
+
+
+def is_equal(arg):
+    return unittest.TestCase.assertGreaterEqual, arg
+
+
+def is_false():
+    return (unittest.TestCase.assertFalse,)
